+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- .NET 7 added as a target framework
+
+### Removed
+
+- .NET Core 3.1 removed as a target framework
+- .NET 5.0 removed as a target framework
+
 ## [2.5.2] - 2022-11-02
 
 ### Changed