+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.0] - 2022-12-08
+
+### Added
+
+- DelayedStateMonitor for IState where onStateLeft returns a fault context that is passed to onStateReached
+
 ## [2.6.0] - 2022-12-06
 
 ### Added